requirements.txt
websocket-client
requests
tls-client
pypasser
names
colorama
curl_cffi
aiohttp
flask
flask_cors
streamlit
selenium
fake-useragent
twocaptcha
pydantic
pymailtm
Levenshtein
retrying
mailgw_temporary_email
pycryptodome
random-password-generator
numpy>=1.22.2 # pinned to avoid a vulnerability
tornado>=6.3.2 # pinned to avoid a vulnerability
PyExecJS
browser_cookie3
googletrans==4.0.0rc1
gevent
openai
python-dotenv
